Match-id-6ad9413eb86535c015c02b6160f3749a27cf0f0c

+++ src/ProtectManager/component/PM_System_Base_Common_Service/src/pm-framework/pm-access-framework/pm-protection-access-framework/src/main/java/openbackup/data/access/framework/copy/mng/service/impl/CopyServiceImpl.java
@@ -677,14 +677,16 @@ public class CopyServiceImpl implements CopyService, CopyServiceSdk {
     }
 
     @Override
-    public void deleteInvalidCopies(String sourceId, int limit) {
+    public void deleteInvalidCopies(String sourceId, int limit, List<String> excludeCopies) {
         BasePage<Copy> checkPointInvalidCopies = queryCopiesByResourceIdAndStatusAndExtendType(sourceId,
             CopyStatus.INVALID.getValue(), CopyExtendType.CHECKPOINT.getValue(), limit);
         List<Copy> deleteCopies = checkPointInvalidCopies.getItems();
         for (Copy c : deleteCopies) {
-            log.info("delete checkPoint invalid copies, total_num:{}, copy_id:{}", checkPointInvalidCopies.getTotal(),
-                c.getUuid());
-            copyRestApi.deleteCopy(c.getUuid(), null);
+            if (!excludeCopies.contains(c.getUuid())) {
+                log.info("delete checkPoint invalid copies, total_num:{}, copy_id:{}",
+                    checkPointInvalidCopies.getTotal(), c.getUuid());
+                copyRestApi.deleteCopy(c.getUuid(), null);
+            }
         }
     }
